Refactor: Split columnar filter.rs into smaller modules
Split the 912-line filter.rs module into focused sub-modules for better
maintainability and code organization:

- filter/comparison.rs - Value comparison logic (compare_values)
- filter/evaluation.rs - Predicate evaluation (evaluate_predicate_tree, evaluate_predicate)
- filter/predicates.rs - Data structures and extraction (PredicateTree, ColumnPredicate, extract_*)
- filter/mod.rs - Main public API and re-exports

Benefits:
- Improved code readability with clear separation of concerns
- Easier to navigate and maintain specific functionality
- Reduced cognitive load when working on filter logic
- Public API remains completely unchanged

Testing:
- All 6 filter tests pass
- All 55 columnar module tests pass
- No performance regression

Closes #2459
